% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHeight.R
\name{checkHeight}
\alias{checkHeight}
\title{Compare the height of persons in a dataset with the mean of the height of the other persons}
\usage{
checkHeight(students.input, sex.specific = TRUE, print.statement = FALSE)
}
\arguments{
\item{students.input}{The original students data.frame}

\item{sex.specific}{Sex specific mean or mean of population}

\item{print.statement}{Print a message after the calculation}
}
\value{
Return the newly created data.frame, difference in height to gender specific mean
}
\description{
Compare the height of persons in a dataset with the mean of the height of the other persons
}
