% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean}
\alias{mean}
\title{Mean value of a vector}
\usage{
mean(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the mean of a \code{(n x 1)} - vector as a scalar value
}
\description{
Mean value of a vector
}
\examples{
x = c(2,2,2)
mean(x)
# [1] 2

}
\author{
Darya Busen <\email{dasha-89.89@mail.ru}>

Return the rounded mean
}
