% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\docType{data}
\name{students}
\alias{students}
\title{The information of students}
\format{A data frame with 8 rows and 5 variables:
\describe{
  \item{age}{age of the students}
  \item{weight}{weight of the students, in kg}
  \item{height}{height of the student, in m}
  \item{sex}{sex of the students: (F) or (M)}
  \item{name}{name of the students}
  }}
\usage{
students
}
\description{
A dataset containing the age, weight, height, sex, name
}
\keyword{datasets}
