% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHeight.R
\name{checkHeight}
\alias{checkHeight}
\title{Create a dataframe with names and difference in height from sex specific mean}
\usage{
checkHeight(students.input = students)
}
\arguments{
\item{students.input}{[\code{data.frame}]\cr
The original students dataframe
Default is students}
}
\value{
[\code{data.frame}]\cr
  Return the newly created data.frame
}
\description{
Create a dataframe with names and difference in height from sex specific mean
}
