% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\docType{data}
\name{students}
\alias{students}
\title{The information of students}
\format{A data frame with 8 rows and 5 variables:
\describe{
  \item{name}{name of the students}
  \item{sex}{sex of the students: (F) or (M)}
  \item{height}{height of the student, in m}
  ...
  }}
\usage{
students
}
\description{
A dataset containing the name, sex, age, height...
}
\keyword{datasets}
